feature: Use internal Link component
Since Gatsby has it's own Link component to route to internal
pages which is a lot faster than the simple `a` tag, this
is now used. Also checking for external links with a
simple check for `http` or ´https` at the start of
the URL. This should do the trick for most of the
URLs.

Signed-off-by: Niklas Metje <22395665+niklasmtj@users.noreply.github.com>